leopold: Move fc980c into hasu subdir to prepare for addition of new controller config
