+++ data/mappings/keyboard_aliases.hjson
@@ -207,7 +207,7 @@
         "target": "leopold/fc660c"
     },
     "fc980c": {
-        "target": "leopold/fc980c"
+        "target": "leopold/fc980c/hasu"
     },
     "gh60": {
         "target": "gh60/revc"
