+++ keyboards/leopold/fc980c/fc980c.c
@@ -1,33 +1,18 @@
-/*
-Copyright 2017 Balz Guenat
+// Copyright 2024 Wilhelm Schuster
+// Copyright 2017 Balz Guenat
+// SPDX-License-Identifier: GPL-2.0-or-later
 
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include QMK_KEYBOARD_H
+#include "fc980c.h"
 
 #include "debug.h"
 #include "i2c_master.h"
 
 keyboard_config_t fc980c_config;
 
-void matrix_init_kb(void) {
+void fc980c_matrix_init_kb(void) {
 #ifdef ACTUATION_DEPTH_ADJUSTMENT
     i2c_init();
 #endif
-
-    matrix_init_user();
 }
 
 void fc980c_eeconfig_update_kb(void) {
@@ -37,16 +22,14 @@ void fc980c_eeconfig_update_kb(void) {
 #endif
 }
 
-void eeconfig_init_kb(void) {
+void fc980c_eeconfig_init_kb(void) {
 #ifdef ACTUATION_DEPTH_ADJUSTMENT
     fc980c_config.actuation_point_value = -1;
     fc980c_eeconfig_update_kb();
 #endif
-
-    eeconfig_init_user();
 }
 
-void keyboard_post_init_kb(void) {
+void fc980c_keyboard_post_init_kb(void) {
 #ifdef ACTUATION_DEPTH_ADJUSTMENT
     fc980c_config.raw = eeconfig_read_kb();
 
@@ -54,6 +37,4 @@ void keyboard_post_init_kb(void) {
         ad5258_write_rdac(fc980c_config.actuation_point_value);
     }
 #endif
-
-     keyboard_post_init_user();
 }
