+++ keyboards/leopold/fc980c/fc980c.h
@@ -1,19 +1,6 @@
-/*
-Copyright 2017 Balz Guenat
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
+// Copyright 2024 Wilhelm Schuster
+// Copyright 2017 Balz Guenat
+// SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
 
@@ -31,11 +18,10 @@ typedef union {
     };
 } keyboard_config_t;
 
-void matrix_init_kb(void);
+void fc980c_matrix_init_kb(void);
 
 void fc980c_eeconfig_update_kb(void);
 
-void eeconfig_init_kb(void);
+void fc980c_eeconfig_init_kb(void);
 
-void keyboard_post_init_kb(void);
-
+void fc980c_keyboard_post_init_kb(void);
