+++ keyboards/leopold/fc980c/hasu/keyboard.json
