+++ keyboards/leopold/fc980c/hasu/rules.mk
@@ -4,4 +4,4 @@
 I2C_DRIVER_REQUIRED = yes
 
 CUSTOM_MATRIX = yes
-SRC += matrix.c ad5258.c actuation_point.c
+SRC += matrix.c keyboards/leopold/ad5258.c keyboards/leopold/actuation_point.c
